Add appliance assets and implement cost estimation functionality

+++ lib/models/consumption_data.dart
@@ -1,6 +1,6 @@
 class ConsumptionData {
   final String day;
-  final int usage;
+  final double usage;
 
   ConsumptionData({required this.day, required this.usage});
 }
